Add lower left resize handle, and some reordering.
An attempt at writing one function that can position all resize handles
may not work, because it relies on the sign of the calculations of the
difference between the corner of the viewport and the opposite corner of
the resize handle.  But this sign can actually differ for the same
handle/corner as the user drags, because the user can drag beyond the
opposite corner, which will flip the sign (and we think this actually
means it's a different corner).

Also I reordered code that worked on all 4 corners to always do it in
the following order: UL, UR, LR, LL (i.e. clockwise)