Refactoring + Checkpoint on moving resize handle code to it's own class.
Refactored some functions from Croptastic to the global scope, since
they're being used in both Croptastic and CroptasticResizeHandle.

Additionally, progress on the new resize handle class: The upper right
handle is drawn with the new code and the remaining handles are using
the old code.  The handle works except for one particular case where the
calculations send the resize handle off to a y-coordinate of negative
infinity.  It happens when the user drags up and then down very quickly
and I think it's because the upper right corner of the viewport gets
very close (in y-coordinate) to the lower-left corner of the resize
handle.