First part of getting rid of inward_positive_{x,y}
It's a complicated and unnecessary notion.  We can separate the
potential handle positions into groups and use membership in those
groups to decide the sign of calculations.  It should be a lot clearer
to read and easier to understand when it's done.

Also, there is a lot of duplicated logic inside drawResizeHandle and
positionHandle that needs to be eliminated.