Rewrite drag to use simpler metric for calculating new viewport size.
Now we set dx/dy as the distance of the mouse cursor from the resize
handle. Much simpler and lets us eliminate most uses of the "fixed
point" concept, which wasn't really elegant, anyway.  We can rename the
remaining uses of "fixed point" to "origin" to be more consistent with
Raphael, since now it's only used as the origin for the scale operation
on the viewport.